Resolve #18 - standardize included file names
Shortcodes were using `css_extern` for adding code in `<head>`, while
sidebar widgets used `css`, and both used `js` for things to put at the
end of `<body>`. These have been changed, so it is now `head` and
`foot`, and `css` is used to add CSS to the generated CSS file.