Foundation and modularization
First off, this commit replaces the homegrown responsive framework with
Foundation by Zurb. This is largely due to wanting a framework that took
care of a few other things and offloading a bit of work so I can focus
on other things - after this refactoring, of course.

Using Foundation also means that I have to do less documentation for
developing plugins for this theme, which is the next big change:
shortcodes, sidebar widgets, and comment providers are all considered
plugins by the theme, which means users can define their own and
integrate them into the website (almost) like it was built into the
theme.

Features I haven't used in a long time and can't easily touch were moved
out of the theme and into an "archive" overlay - in case anyone out
there is using those features, i.e. HashOver or events integration. The
archive theme overlay can be used with Hugo's theme composition to
include the features in the site despite being in a different theme.

Finally, support for Commento self-hosted comments has been added to
BluestNight as the default comments provider. I would have kept Muut,
but they seem to have gotten rid of the free plan I seem to remember
them having. I needed a usable service to test the comment provider
plugin system with, so Commento was added in this commit.

Various small bug fixes were added as part of refactoring the code base
to use Foundation, as well as changes like new partials for determining
whether to use a minified version of a resource and using the
"make_link" partial everywhere instead of the various built-in
*(Lang)URL functions - mostly for consistency and in case I want to
change the default link-making behavior.

I have hand-tested these changes pretty thoroughly, but there may be
things I missed. There is still the issue of more generic plugins before
this gets merged for release, though, and that will take more testing.