Add support for generic plugins
Generic plugins add only CSS or JavaScript to a page and are added to
every page. They are meant to be used for things like CSS customization
or adding JavaScript to every page (like highlight.js, maybe?).

Code isn't tested because I have no use for it right now, but it should
work given how similar it is to the other kinds of plugins.