Fix loading of Content with shortcodes in headers
Hugo - for some reason - adds a string like
"HAHAHUGOSHORTCODE-0x349582325-3-HBHB" when a shortcode is used in a
header. This interferes with me using the `fa` shortcode in the headers,
so I had to do some fun regular expression stuff to remove that junk
from the table of contents and the generated ids/links.

The one problem caused by this is if someone creates two headers with
the same text and use shortcodes in the prefix - each "HAHA" text is
unique, so both headers will have the same id and the same link, so both
entries in the TOC will link to the first instance.