CSS Improvements
Decided I would rename some of the color nicknames in the CSS, since the
names they had previously didn't make clear which *text color went with
which *bg colors.

Also fixed the loading of the background image with relative links.
Apparently CSS variables containing paths are resolved relative to the
file the variable is referenced in, rather than the one it is set in. So
I had to tweak the `make_link` partial a little to get it to generate
the correct link.