Resolve #6 - confirm form submission above button
Users with the submit button at the bottom of the page might not see the
"Form submitted" confirmation. This led to a user submitting the form
multiple times, only to realize that the confirmation was hidden.

The confirmation now appears above the button, in such a way that it
should always be visible if the button was visible when clicked.