Resolve #9 - Broken relative URLs on 404
Relative links were broken on the 404 page, because URLs were generated
according to the path `/404.html`, while the broken path might have a
different number of subdirectories. The only real way to work around it
is to force the 404 page to use root-relative links.

Also bumped the minimum required Hugo version because of using variable
reassignment instead of `Scratch`.