Resolve #8 - Add author byline on list page
When multiple authors exist on a site, Midnight didn't indicate who
wrote what on the list pages. This fixes that by adding a byline if
there are multiple members defined on the site. It also checks that the
author exists as a member in the site's data. If the author is not
defined for a page, the site default is used.