Fix async CSS
I had originally attempted to make the `link[rel=preload]` tags change
type to `stylesheet` by binding event handlers in JavaScript, so I could
remove all inline JS for #14. Most of the files were finished
downloading by the time that happened, though, so the event handlers
never fired.

This uses the actual `loadCSS` from the loadCSS project that provides
the `rel=preload` compatibility script, which creates an element in
JavaScript that loads the file and, when the file "finishes" downloading
(or validates in the cache), replaces the CSS entry in the `<head>` to
set `stylesheet`. All without any inline scripts on the page.