Improve Iubenda loading
Replace dynamic insertion of a script tag with the script tag itself (at
the end of the body) and replace the cookie banner JS file with the
versioned and cacheable one that users will get redirected to. The only
downside of this is having to manually update that URL when new versions
are released.