enable_cookie_policy -> enable_cookie_banner
Iubenda's cookie policy is separate from their cookie banner solution.
The configuration option `enable_cookie_policy` was being used for the
cookie banner, which is confusing now that Midnight includes support for
the actual cookie policy.

This commit replaces references to `enable_cookie_policy` in the context
of the cookie banner to `enable_cookie_banner`.