Fix loading of search index on relative sites.
Sites configured with relative links and subdomain base URLs failed to
load the search index `index.json`. This fix requires some JavaScript
code to determine what relative URL to use to download the index file.

Sites not using relative links will not use the above code, since the
path is the same for all pages.

This commit also fixes an issue with generating absolute urls on sites
with subdirectories in the baseURL.