Prepare for recursive configuration
I've submitted a pull request for Hugo to recursively merge
configurations and it looks like it will be included more or less as is
(maybe with a command flag or something to enable it).

This removes the "Midnight" namespacing of everything that was required
previously - the module configuration files now set everything in the
exact location users will. This has been surface-level tested to work.